% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_linearization_functions.R
\name{compute_allocation_prob}
\alias{compute_allocation_prob}
\title{Computing the component allocation probabilities}
\usage{
compute_allocation_prob(y, linear_pred, n_comp)
}
\arguments{
\item{y}{A vector of response variable values.}

\item{linear_pred}{A vector of the linear predictors in both components and mixture weights models.}

\item{n_comp}{Number of mixture components.}
}
\value{
The components allocation probabilities.
}
\description{
Computing the component allocation probabilities
}
