% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{design_matrix}
\alias{design_matrix}
\title{function for preparing the design matrix for mixture of experts models}
\usage{
design_matrix(x, z, n_comp)
}
\arguments{
\item{x}{A vector of covariates for components models.
The vector doesn't include the intercept.}

\item{z}{A vector of covariates for mixture weights models.
The vector doesn't include the intercept.}

\item{n_comp}{Number of mixture components.}
}
\value{
The design matrix.
}
\description{
function for preparing the design matrix for mixture of experts models
}
