% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_density_functions.R, R/gp_particle_filter.R
\name{dgenpois}
\alias{dgenpois}
\title{density function of the generalized Poisson}
\usage{
dgenpois(
  train_data,
  intervals,
  n_particles,
  train_col,
  time_ind = 1,
  response_ind = 2,
  alpha = 0.5,
  r = 2,
  return_allweights = F
)

dgenpois(
  train_data,
  intervals,
  n_particles,
  train_col,
  time_ind = 1,
  response_ind = 2,
  alpha = 0.5,
  r = 2,
  return_allweights = F
)
}
\arguments{
\item{y}{: the response variable.}

\item{mu}{: is the mean parameter for the generalized poisson distribution}

\item{lambda}{: is the scale parameter describing the dispersion around the mean.}

\item{log_scale:}{is a logic indicating whether the density value returned should be in log scale or not}

\item{train_data:}{a data frame containing the training data set.}

\item{intervals:}{a vector of subsequent time points partitioning the training period.}

\item{n_particles:}{number of parameters to be sampled (posterior sample size)}

\item{train_col:}{a vector of column names that are used as covariates. These names should be consistent with the column names in the train_data}

\item{time_ind:}{column index of the up dates.}

\item{response_ind:}{column index of the response variable}

\item{alpha:}{discount factor; it controls the smoothness of the parameters across intervals.}

\item{r:}{number of replicates each sampled parameter appears in the posterior sample.
This is used in the resampling step of the particle filter (see file stratified_rejection_sampling.R)}

\item{return_allweights:}{a logic value that indicates if all weights of the sampled posterior parameters should be returned}
}
\value{
density value

:
particles_mean, particles_disp: a sample from the posterior
ESS: Effective sample size
weights: importance weights
lps: log predictive score
}
\description{
dgenpois computes the density of a given count value (y).

This function implements the particle filter algorithm for the dynamic generalized Poisson regression model.
}
