% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_particle_filter.R
\name{dgp}
\alias{dgp}
\title{particle filter algorithm for the dynamic generalized Poisson model}
\usage{
dgp(
  train_data,
  intervals,
  n_particles,
  train_col,
  time_ind = 1,
  response_ind = 2,
  alpha = 0.5,
  r = 2,
  h = 1,
  return_allweights = F
)
}
\arguments{
\item{train_data:}{a data frame containing the training data set.}

\item{intervals:}{a vector of subsequent time points partitioning the training period.}

\item{n_particles:}{number of parameters to be sampled (posterior sample size)}

\item{train_col:}{a vector of column names that are used as covariates. These names should be consistent with the column names in the train_data}

\item{time_ind:}{column index of the up dates.}

\item{response_ind:}{column index of the response variable}

\item{alpha:}{discount factor; it controls the smoothness of the parameters accross intervals.}

\item{r:}{number of replicates each sampled parameter appears in the posterior sample.
This is used in the resampling step of the particle filter (see file stratified_rejection_sampling.R)}

\item{return_allweights:}{a logic value that indicates if all weights of the sampled posterior parameters should be returned}
}
\value{
:
particles_mean, particles_disp: a sample from the posterior
ESS: Effective sample size
weights: importance weights
lps: log predictive score
}
\description{
This function implements the particle filter algorithm for the dynamic generalized Poisson regression model.
}
