% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moe_particle_filter.R
\name{dmoe}
\alias{dmoe}
\title{particle filter algorithm for dynamic mixture of experts}
\usage{
dmoe(
  Data,
  intervals,
  particle_size,
  exp_col,
  mix_col = NULL,
  n_comp = 1,
  v_init = 1,
  time_ind = 1,
  response_ind = 2,
  alpha = 0.45,
  duplicate_num = 2,
  proposal_method = "linearbayes",
  return_all = F
)
}
\description{
particle filter algorithm for dynamic mixture of experts
}
