% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_processes.R
\name{dynamic_poisson}
\alias{dynamic_poisson}
\title{Simulate data from single component dynamic Poisson data generating process}
\usage{
dynamic_poisson(parameter, dim.design, dates, n)
}
\description{
Simulate data from single component dynamic Poisson data generating process
}
