% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_processes.R
\name{dynamic_poisson_mixture}
\alias{dynamic_poisson_mixture}
\title{Simulate data from dynamic Poisson mixture of experts data generating process}
\usage{
dynamic_poisson_mixture(beta, theta, dim.design, dates, n)
}
\description{
Simulate data from dynamic Poisson mixture of experts data generating process
}
