% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_weights.R
\name{gp_computeweight}
\alias{gp_computeweight}
\title{function for computing importance weights for dynamic generalized Poisson model.}
\usage{
gp_computeweight(
  y,
  x,
  prev_weights,
  beta,
  theta,
  prev_beta,
  prev_theta,
  w_beta,
  w_theta,
  proposalmean_mean,
  proposalcov_mean,
  proposalmean_disp,
  proposalcov_disp
)
}
\arguments{
\item{y:}{a vector of the response values observed in the batch j.}

\item{x:}{a matrix of the covariates observed in the batch j.}

\item{prev_weights:}{weights of the parameters in the previous intervals j-1.}

\item{beta:}{a vector of proposed regression parameters on the mean function}

\item{theta:}{a vector of proposed regression parameters on the overdispersion (scale) function}

\item{prev_beta:}{a vector of regression parameters on the mean function in the previous interval j-1.}

\item{prev_theta:}{a vector of regression parameters on the overdispersion (scale) function in the previous interval j-1.}

\item{w_beta:}{transition matrix defining the evolution of beta from prev_beta.}

\item{w_theta:}{transition matrix defining the evolution of theta from prev_theta.}

\item{proposalmean_mean:}{a vector of the mean value of the proposal distribution of the regression parameters on the mean function.}

\item{proposalmean_disp:}{a vector of the mean value of the proposal distribution of the regression parameters on the overdispersion(scale) function.}

\item{proposalcov_mean:}{covariance matrix of the proposal distribution of the regression parameters on the mean function.}

\item{proposalcov_disp:}{covariance matrix of the proposal distribution of the regression parameters on the overdispersion(scale) function.}
}
\value{
updated importance weights.
}
\description{
This function computes the weights of the sampled parameters.
The weights are functions of the likelihood and the prior.
}
