% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_density_functions.R
\name{gp_likelihood}
\alias{gp_likelihood}
\title{generalized Poisson likelihood function
dgp_likelihood computes the likelihood value of a response variable y given the covariate x.}
\usage{
gp_likelihood(y, x, beta, theta)
}
\arguments{
\item{y:}{a vector of the response values}

\item{x:}{a matrix of the covariate values. the matrix should have the rows = length(y) and columns= number of covariates+1}

\item{beta:}{regression coefficients on the mean function of the generalized poisson model}

\item{theta:}{the regression coefficients on the scale (overdispersion) parameter of the generalized poisson model}
}
\value{
log_like: vector of the loglikelihood values for y
}
\description{
generalized Poisson likelihood function
dgp_likelihood computes the likelihood value of a response variable y given the covariate x.
}
