% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_moments.R
\name{gp_linearbayes}
\alias{gp_linearbayes}
\title{proposal for the generalized Poisson component model}
\usage{
gp_linearbayes(y, design_matrix, mean_effect, disp_effect, u_mean, u_disp)
}
\arguments{
\item{y:}{a vecotor of the response values}

\item{design_matrix:}{a matrix of the covariate values augmented by a column of 1s}

\item{mean_effect:}{current value of the mean function of the generalized poisson model}

\item{disp_effect:}{current value of the overdispersion (scale) function of the generalized poisson model}

\item{u_mean:}{transition matrix of the regression parameters on the mean function}

\item{u_disp:}{transition matrix of the regression parameters on the overdispersion function.}
}
\value{
: updated mean_effect, disp_effect, cov_mean and cov_disp.
}
\description{
this function computes the moments (mean and covariance) of the proposal distribution
}
