% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_density_functions.R
\name{gp_pred_dens}
\alias{gp_pred_dens}
\title{generalized Poisson predictive density function
the following function computes the predictive distribution for the generalized Poisson model}
\usage{
gp_pred_dens(test_data, beta, theta, importance_weight, y_max, training_col)
}
\arguments{
\item{test_data:}{test data}

\item{beta:}{regression parameter on the mean function}

\item{theta:}{regression parameter on the overdispersion parameter.}

\item{importance_weight:}{weight of the parameter.}

\item{y_max:}{maximum value of the response variable.}

\item{training_col:}{a vector of the names of variables used in the training.}
}
\value{
empirical predictive distribution
}
\description{
generalized Poisson predictive density function
the following function computes the predictive distribution for the generalized Poisson model
}
