% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_linearization_functions.R
\name{gradient_posterior}
\alias{gradient_posterior}
\title{Posterior gradients.}
\usage{
gradient_posterior(allocation_prob, y, linear_pred, n_comp)
}
\arguments{
\item{allocation_prob}{A vector of component allocation probabilities.}

\item{y}{A vector of response variable values.}

\item{linear_pred}{A vector of the linear predictors in both components and mixture weights models.}

\item{n_comp}{Number of mixture components.}
}
\value{
A vector of posterior gradients.
}
\description{
Computing the posterior gradients of the linear predictors.
}
