% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_linearization_functions.R
\name{hessian_posterior}
\alias{hessian_posterior}
\title{Posterior hessian.}
\usage{
hessian_posterior(
  allocation_prob,
  posterior_gradients,
  prior_var_inv,
  y,
  linear_pred,
  n_comp
)
}
\arguments{
\item{allocation_prob}{A vector of component allocation probabilities.}

\item{posterior_gradients}{A vector of posterior gradients of the linear predictors.}

\item{prior_var_inv}{Inverse of the covariance matrix of the prior for the regression coefficients.}

\item{y}{A vector of response variable values.}

\item{linear_pred}{A vector of the linear predictors in both components and mixture weights models.}

\item{n_comp}{Number of mixture components.}
}
\value{
The posterior hessian matrix.
}
\description{
Computing the posterior gradients with respect to the linear predictors.
}
