% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance_weights.R
\name{importance_weight_multicomp}
\alias{importance_weight_multicomp}
\title{function for computing importance weights for the dynaminc mixture of experts model.}
\usage{
importance_weight_multicomp(
  y,
  x,
  z,
  param,
  prev_weights,
  prev_particles,
  prior_mean,
  prior_var,
  proposal_mean,
  proposal_cov,
  n_comp
)
}
\arguments{
\item{y:}{a vector of the response values observed in the batch j.}

\item{x:}{a matrix of the covariates observed in the batch j.}

\item{prev_weights:}{weights of the parameters in the previous intervals j-1.}

\item{param:}{A vector of regression coefficeints (one particle sample).}

\item{prev_particles:}{Posterior sample for the previous interval j-1.}

\item{prior_mean:}{Prior mean for the regression coefficients.}

\item{prior_var:}{Prior covariance matrix for regression coefficients.}

\item{proposal_mean:}{Proposal mean vector for the regression coefficients.}

\item{proposal_cov:}{Proposal covariance matrix for the regression coefficients.}

\item{n_comp:}{Number of mixture components.}
}
\value{
updated importance weights.
}
\description{
This function computes the weights of the sampled parameters.
The weights are functions of the likelihood and the prior.
}
