% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_moments.R
\name{local_linear_proposal_1comp}
\alias{local_linear_proposal_1comp}
\title{Compute the updated first and second moments for the proposal density for
 single component models using local linearization of the target online posterior density.}
\usage{
local_linear_proposal_1comp(y, x, reg_coef, cov_reg_coef)
}
\arguments{
\item{y}{A vector of response variable values.}

\item{x}{A matrix of dimension h x p containing the covariates data,
where h is the number of observations (equal to the length of y) and
p represent the dimension the covariates in the component models.}

\item{reg_coef}{A vector of the regression coefficients.}

\item{cov_reg_coef}{The covariance matrix of the regression coefficients.}
}
\value{
Updated mean and covariance matrix of the regression parameter.
}
\description{
Compute the updated first and second moments for the proposal density for
 single component models using local linearization of the target online posterior density.
}
