% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_moments.R
\name{local_linear_proposal_multicomp}
\alias{local_linear_proposal_multicomp}
\title{Proposal function based on the local linearization of the target online posterior density.}
\usage{
local_linear_proposal_multicomp(y, x, z, reg_coef, cov_reg_coef, n_comp)
}
\arguments{
\item{y}{A vector of response variable values.}

\item{x}{A matrix of dimension h x p containing the covariates data,
where h is the number of observations (equal to the length of y) and
p represent the dimension the covariates in the component models.}

\item{z}{A matrix of dimension h x q containing the covariates data,
where h is the number of observations (equal to the length of y) and
q represent the dimension the covariates in the mixture weights model.}

\item{reg_coef}{A vector of the regression coefficients in both
component and mixture weight models.}

\item{cov_reg_coef}{The prior covariance matrix of the regression coefficients
in both component and mixture weight models.}

\item{n_comp}{Number of mixture components.}
}
\value{
Updated mean and covariance matrix of the regression coefficients.
}
\description{
Computes the updated first and second moments for the proposal density for
 mixture models using local linearization of the target online posterior density.
}
