% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_density_functions.R
\name{lps_gpoisson}
\alias{lps_gpoisson}
\title{Compute LPS for generized Poisson model}
\usage{
lps_gpoisson(x, y, particles_beta, particles_theta, importance_weight)
}
\arguments{
\item{y}{: a vector of the response values.}

\item{x:}{a matrix of the covariate values. the matrix should have the rows = length(y) and columns= number of covariates+1}

\item{particles_beta:}{a posterior sampled of the regression cofficients on the mean function}

\item{particles_theta:}{a posterior sample of the regression coefficients on the overdispersion function}

\item{importance_weight:}{a vector of the weights of sampled parameters.}
}
\value{
lps value
}
\description{
the function lps_gpoisson computes the log predictive score: a measure used to compare models
}
