% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{mixture_log_likelihood}
\alias{mixture_log_likelihood}
\title{Log likelihood for the mixture of experts model}
\usage{
mixture_log_likelihood(y, x, z, param, n_comp)
}
\description{
Log likelihood for the mixture of experts model
}
