% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_rejection_resampling.R
\name{normalize}
\alias{normalize}

\title{A function for normalizing any set of empirical densities.}

\usage{
normalize(values)
}
\arguments{
\item{values}{A vector of unnormalized densities}
}
\value{
A vector of normalized densities
}
\description{

A function for normalizing any set of empirical densities.

}
