% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_rejection_resampling.R
\name{partition}
\alias{partition}
\title{Partition}
\usage{
partition(w, n)
}
\arguments{
\item{w}{A vector of importance weights (of length M).}

\item{n}{The desired size of output vector (N < M).}
}
\value{
The threshold of the largest value that is less than all the importance weights of the n best particles.
}
\description{
The partition function implements the partition algorithm used to search particles that should be rejected/ discarded
It is based on the cumulative distribution of the importance weights.
}
