% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_linearization_functions.R
\name{posterior_linear_moments}
\alias{posterior_linear_moments}
\title{Local linear approximation of the posterior.}
\usage{
posterior_linear_moments(y, linear_pred, prior_var_inv, n_comp)
}
\arguments{
\item{y}{A response value.}

\item{linear_pred}{A vector of the linear predictors in both components and mixture weights models.}

\item{prior_var_inv}{Inverse of the covariance matrix of the prior for the regression coefficients.}

\item{n_comp}{Number of mixture components.}
}
\value{
The the first and second moments of the linear approximation of the posterior.
}
\description{
Computes the first and second moments of the linear approximation of the posterior using Newton's method.
}
