% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{pred_density}
\alias{pred_density}
\title{Predictive density for dynamic mixture of experts.}
\usage{
pred_density(model, test_data, y_max = NULL, batch_ind = NULL)
}
\description{
Predictive density for dynamic mixture of experts.
}
