% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictions_1Comp}
\alias{predictions_1Comp}
\title{Predictive density for dynamic Poisson regression.}
\usage{
predictions_1Comp(test_data, Particles, ImportanceWeight, y.max, exp.col)
}
\description{
Predictive density for dynamic Poisson regression.
}
