% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predictions_mutlicomp}
\alias{predictions_mutlicomp}
\title{Predictive density for dynamic mixture of experts.}
\usage{
predictions_mutlicomp(
  test_data,
  particles,
  importance_weight,
  y_max,
  exp_col,
  mix_col,
  n_comp
)
}
\description{
Predictive density for dynamic mixture of experts.
}
