% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_rejection_resampling.R
\name{reject_strat_resample}
\alias{reject_strat_resample}
\title{The following function implements the stratified sampling algorithm with rejection
The resampling scheme is based on Paul Fearnhead and Peter Clifford's paper:
" Online inference for hidden Markov models via particle filters"}
\usage{
reject_strat_resample(w, n)
}
\arguments{
\item{w}{A vector of importance weights (of length M).}

\item{n}{The desired size of output vector (N < M).}
}
\value{
A vector of selected indexes of length n.
}
\description{
The following function implements the stratified sampling algorithm with rejection
The resampling scheme is based on Paul Fearnhead and Peter Clifford's paper:
" Online inference for hidden Markov models via particle filters"
}
