% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_linearization_functions.R
\name{robust_inverse}
\alias{robust_inverse}
\title{robust_inverse}
\usage{
robust_inverse(mat)
}
\arguments{
\item{mat}{the covariance matrix}
}
\value{
The inverse of the covariance matrix.
}
\description{
Compute the inverse of a covariance matrix.
}
