% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_moments.R
\name{single_component_proposal}
\alias{single_component_proposal}
\title{Proposal function's moments for Single component models}
\usage{
single_component_proposal(y, DesignMatrix, effect.param, cov_param)
}
\arguments{
\item{y}{A vector of response variable values.}

\item{effect.param}{The prior mean of the regression parameter.}

\item{cov_param}{The prior covariance matrix of the regression parameter.}

\item{DesingMatrix}{A matrix of dimension h x (p+1) containing the covariates data, where h is the number of observations (equal to the length of y) and p represent the dimension the covariates.
Note that the the first column is a column  of 1s which is added in order to allow for the intercept in the model.}
}
\value{
Updated mean and covariance matrix of the regression parameter.
}
\description{
Computes the updated first and second moments for the proposal density of single component models.
}
