% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_generating_processes.R
\name{static_poisson}
\alias{static_poisson}
\title{Simulate data from static Poisson data generating process}
\usage{
static_poisson(parameter, dim.design, dates, n)
}
\description{
Simulate data from static Poisson data generating process
}
