% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_rejection_resampling.R
\name{strat_resample}
\alias{strat_resample}
\title{implementation of the stratified sampling algorithm.}
\usage{
strat_resample(w, n)
}
\arguments{
\item{w}{A vector of importance weights (of length M).}

\item{n}{The desired size of output vector (N < M).}
}
\value{
A vector of selected indexes of length n.
}
\description{
implementation of the stratified sampling algorithm.
}
