% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_linearization_functions.R
\name{update_linear_moments}
\alias{update_linear_moments}
\title{Update the first and second moments of regression coefficients given
their hessian  and the gradient.}
\usage{
update_linear_moments(prior_mean, inv_prior_var, gradient, hessian_matrix)
}
\arguments{
\item{prior_mean}{Prior mean of the regression coefficients.}

\item{inv_prior_var}{Inverse of the covariance matrix of the regression coefficients.}

\item{gradient}{The gradient of the regression coefficients.}

\item{hessian_matrix}{The Hessian matrix of the regression coefficients.}
}
\value{
The updated location and covariance matrix of the regression coefficients.
}
\description{
Update the first and second moments of regression coefficients given
their hessian  and the gradient.
}
