% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gp_density_functions.R
\name{dgenpois}
\alias{dgenpois}
\title{density function of the generalized Poisson}
\usage{
dgenpois(y, mu, lambda, log_scale = FALSE)
}
\arguments{
\item{y}{: the response variable.}

\item{mu}{: is the mean parameter for the generalized poisson distribution}

\item{lambda}{: is the scale parameter describing the dispersion around the mean.}

\item{log_scale:}{is a logic indicating whether the density value returned should be in log scale or not}
}
\value{
density value
}
\description{
dgenpois computes the density of a given count value (y).
}
