% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_moments.R
\name{linear_bayes_proposal_1comp}
\alias{linear_bayes_proposal_1comp}
\title{Compute the updated first and second moments for the proposal density of
 single component models using linear Bayes approximations.}
\usage{
linear_bayes_proposal_1comp(y, x, reg_coef, cov_reg_coef)
}
\arguments{
\item{y}{A vector of response variable values.}

\item{x}{A matrix of dimension h x p containing the covariates data,
where h is the number of observations (equal to the length of y) and
p represent the dimension the covariates.}

\item{reg_coef}{A vector of the regression coefficients.}

\item{cov_reg_coef}{The covariance matrix of the regression coefficient.}
}
\value{
Updated mean and covariance matrix of the regression parameter.
}
\description{
Compute the updated first and second moments for the proposal density of
 single component models using linear Bayes approximations.
}
