% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proposal_moments.R
\name{local_linear_proposal_1comp}
\alias{local_linear_proposal_1comp}
\title{Proposal for Single component models using local linearization of the target online posterior density.}
\usage{
local_linear_proposal_1comp(y, x, param, cov_param)
}
\arguments{
\item{y}{A vector of response variable values.}

\item{x}{A matrix of dimension h x (p+1) containing the covariates data, where h is the number of observations (equal to the length of y) and p represent the dimension the covariates in the component models.
Note that the the first column is a column  of 1s which is added in order to allow for the intercept in the model.}

\item{param}{The prior mean of the regression parameter in both component and mixture weight models.}

\item{cov_param}{The prior covariance matrix of the regression parameter in both component and mixture weight models.}

\item{n_comp}{Number of mixture components.}
}
\value{
Updated mean and covariance matrix of the regression parameter.
}
\description{
Computes the updated first and second moments for the proposal density for
 single component models using local linearization of the target online posterior density.
}
