% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_functions.R
\name{poisson_log_likelihood}
\alias{poisson_log_likelihood}
\title{Log likelihood for Poisson regression model}
\usage{
poisson_log_likelihood(y, X, param)
}
\description{
Log likelihood for Poisson regression model
}
